Fixing volatile type

+++ dht.cpp
@@ -176,7 +176,7 @@ int8_t dht::_readSensor(uint8_t pin, uint8_t wakeupDelay, uint8_t leadingZeroBit
     // direct port read is about 3x faster
     uint8_t bit = digitalPinToBitMask(pin);
     uint8_t port = digitalPinToPort(pin);
-    volatile uint8_t *PIR = portInputRegister(port);
+    volatile uint32_t *PIR = portInputRegister(port);
 
     // REQUEST SAMPLE
     pinMode(pin, OUTPUT);
